Support CreateTable constraint and references

+++ src/SQLBuilder/Universal/Syntax/Column.php
@@ -184,7 +184,7 @@ class Column implements ToSqlInterface {
 
     public function int($length = NULL)
     {
-        $this->type = 'mediumint';
+        $this->type = 'integer';
         $this->isa = 'int';
         if ($length) {
             $this->length = $length;
